requirements.txt
attrs
autopep8
backcall
certifi
cftime
charset-normalizer
cycler
decorator
entrypoints
ipykernel
ipython
netCDF4
numpy
pytest
pyaml
PyYAML
requests
StrEnum
